ARM: Towards Zedboard support
- Runs on qemu-system-arm -machine xilinx-zynq-a9
- UART working
- Reset working (used to shutdown qemu -no-reboot)
- Dispatcher working
- Private Timer fires interrupt
- Alarms fire and dispatch
- Interrupts

+++ fail/trace.h
@@ -60,9 +60,9 @@ inlinehint void test_trace(uint32_t val)\
     fail_trace = val;
 
     #ifndef FAIL
-    static uint32_t i = 1;
+	static uint32_t  i = 1;
 
-    kout << "T " << i++;
+    kout << "T " << (int)i++;
     kout << ": " <<  hex << val << dec << endl;
     #endif
 }
